+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/AmazeeLabs/silverback-template/compare/@amazeelabs/silverback-template@2.3.5...@amazeelabs/silverback-template@2.4.0) (2025-08-25)
+
+
+### Bug Fixes
+
+* update types location ([0f471d7](https://github.com/AmazeeLabs/silverback-template/commit/0f471d7d11dbe1df8337356749bea56f93cd3a58))
+
+
+### Features
+
+* add react router bridge ([b3f6d5d](https://github.com/AmazeeLabs/silverback-template/commit/b3f6d5d67f0b528927c7e76a97a03a8f001daa8d))
+
+
+
+
+
 ## [2.3.5](https://github.com/AmazeeLabs/silverback-template/compare/@amazeelabs/silverback-template@2.3.4...@amazeelabs/silverback-template@2.3.5) (2025-08-08)
 
 **Note:** Version bump only for package @amazeelabs/silverback-template